requirements.txt
backcall==0.2.0
decorator==5.1.1
distro==1.9.0
entrypoints==0.4
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
jedi==0.17.2
jupyter-client==7.1.2
jupyter-core==4.9.2
nest-asyncio==1.6.0
numpy==1.19.5
packaging==21.3
parso==0.7.1
pexpect==4.9.0
pickleshare==0.7.5
Pillow==8.4.0
pkg-resources==0.0.0
prompt-toolkit==3.0.36
ptyprocess==0.7.0
Pygments==2.14.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pyzmq==25.1.2
scikit-build==0.16.7
six==1.16.0
tornado==6.1
traitlets==4.3.3
typing-extensions==4.1.1
wcwidth==0.2.13
